Make a couple of methods private and wrap the migration in a transaction
