+++ server/services/store/storetests/blocks.go
@@ -61,16 +61,6 @@ func StoreTestBlocksStore(t *testing.T, setup func(t *testing.T) (store.Store, f
 		defer tearDown()
 		testGetBlock(t, store, container)
 	})
-	t.Run("GetBlocksWithSameID", func(t *testing.T) {
-		store, tearDown := setup(t)
-		defer tearDown()
-		testGetBlocksWithSameID(t, store, container)
-	})
-	t.Run("ReplaceBlockID", func(t *testing.T) {
-		store, tearDown := setup(t)
-		defer tearDown()
-		testReplaceBlockID(t, store, container)
-	})
 }
 
 func testInsertBlock(t *testing.T, store store.Store, container store.Container) {
@@ -673,111 +663,4 @@ func testGetBlock(t *testing.T, store store.Store, container store.Container) {
 		require.NoError(t, err)
 		require.Nil(t, fetchedBlock)
 	})
-}
-
-//nolint:gosec
-func testGetBlocksWithSameID(t *testing.T, storeInstance store.Store, _ store.Container) {
-	container1 := store.Container{WorkspaceID: "1"}
-	container2 := store.Container{WorkspaceID: "2"}
-	container3 := store.Container{WorkspaceID: "3"}
-
-	block1 := model.Block{ID: "block-id-1", RootID: "root-id-1"}
-	block2 := model.Block{ID: "block-id-2", RootID: "root-id-2"}
-	block3 := model.Block{ID: "block-id-3", RootID: "root-id-3"}
-
-	block4 := model.Block{ID: "block-id-1", RootID: "root-id-1"}
-	block5 := model.Block{ID: "block-id-2", RootID: "root-id-2"}
-
-	block6 := model.Block{ID: "block-id-1", RootID: "root-id-1"}
-	block7 := model.Block{ID: "block-id-7", RootID: "root-id-7"}
-	block8 := model.Block{ID: "block-id-8", RootID: "root-id-8"}
-
-	for _, block := range []model.Block{block1, block2, block3} {
-		err := storeInstance.InsertBlock(container1, &block, "user-id")
-		require.NoError(t, err)
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	for _, block := range []model.Block{block4, block5} {
-		err := storeInstance.InsertBlock(container2, &block, "user-id")
-		require.NoError(t, err)
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	for _, block := range []model.Block{block6, block7, block8} {
-		err := storeInstance.InsertBlock(container3, &block, "user-id")
-		require.NoError(t, err)
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	blocksWithDuplicatedID := []model.Block{block1, block2, block4, block5, block6}
-
-	blocks, err := storeInstance.GetBlocksWithSameID()
-	require.NoError(t, err)
-
-	// we process the found blocks to remove extra information and be
-	// able to compare both expected and found sets
-	foundBlocks := []model.Block{}
-	for _, foundBlock := range blocks {
-		foundBlocks = append(foundBlocks, model.Block{ID: foundBlock.ID, RootID: foundBlock.RootID})
-	}
-
-	require.ElementsMatch(t, blocksWithDuplicatedID, foundBlocks)
-}
-
-//nolint:gosec
-func testReplaceBlockID(t *testing.T, storeInstance store.Store, _ store.Container) {
-	container1 := store.Container{WorkspaceID: "1"}
-	container2 := store.Container{WorkspaceID: "2"}
-
-	// blocks from workspace1
-	block1 := model.Block{ID: "block-id-1", RootID: "root-id-1"}
-	block2 := model.Block{ID: "block-id-2", RootID: "root-id-2", ParentID: "block-id-1"}
-	block3 := model.Block{ID: "block-id-3", RootID: "block-id-1"}
-	block4 := model.Block{ID: "block-id-4", RootID: "block-id-2"}
-	block5 := model.Block{ID: "block-id-5", RootID: "block-id-1", ParentID: "block-id-1"}
-
-	// blocks from workspace2. They're identical to blocks 1 and 2,
-	// but they shouldn't change
-	block6 := model.Block{ID: "block-id-1", RootID: "root-id-1"}
-	block7 := model.Block{ID: "block-id-2", RootID: "root-id-2", ParentID: "block-id-1"}
-
-	for _, block := range []model.Block{block1, block2, block3, block4, block5} {
-		err := storeInstance.InsertBlock(container1, &block, "user-id")
-		require.NoError(t, err)
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	for _, block := range []model.Block{block6, block7} {
-		err := storeInstance.InsertBlock(container2, &block, "user-id")
-		require.NoError(t, err)
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	currentID := "block-id-1"
-	newID := "new-id-1"
-	err := storeInstance.ReplaceBlockID(currentID, newID, "1")
-	require.NoError(t, err)
-
-	newBlock1, err := storeInstance.GetBlock(container1, newID)
-	require.NoError(t, err)
-	newBlock2, err := storeInstance.GetBlock(container1, block2.ID)
-	require.NoError(t, err)
-	newBlock3, err := storeInstance.GetBlock(container1, block3.ID)
-	require.NoError(t, err)
-	newBlock5, err := storeInstance.GetBlock(container1, block5.ID)
-	require.NoError(t, err)
-	newBlock6, err := storeInstance.GetBlock(container2, block6.ID)
-	require.NoError(t, err)
-	newBlock7, err := storeInstance.GetBlock(container2, block7.ID)
-	require.NoError(t, err)
-
-	require.Equal(t, newID, newBlock1.ID)
-	require.Equal(t, newID, newBlock2.ParentID)
-	require.Equal(t, newID, newBlock3.RootID)
-	require.Equal(t, newID, newBlock5.RootID)
-	require.Equal(t, newID, newBlock5.ParentID)
-
-	require.Equal(t, currentID, newBlock6.ID)
-	require.Equal(t, currentID, newBlock7.ParentID)
 }
